ref #457 -- organized classes into their own files
