+++ functions/image/timber-image-operation-retina.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * Increases image size by a given factor
+ * Arguments:
+ * - factor by which to multiply image dimensions
+ */
+class TimberImageOperationRetina extends TimberImageOperation {
+
+    private $factor;
+
+    /**
+     * @param int $factor to multiply original dimensions by
+     */
+    function __construct($factor) {
+        $this->factor = $factor;
+    }
+
+    function filename($src_filename, $src_extension) {
+        $newbase = $src_filename . '@' . $this->factor . 'x'; // add @2x, @3x, @1.5x, etc.
+        $new_name = $newbase . '.' . $src_extension;
+        return $new_name;
+    }
+
+    function run($load_filename, $save_filename){
+        $image = wp_get_image_editor( $load_filename );
+        if ( !is_wp_error( $image ) ) {
+            $current_size = $image->get_size();
+            $src_w = $current_size['width'];
+            $src_h = $current_size['height'];
+            $src_ratio = $src_w / $src_h;
+            // Get ratios
+            $w = $src_w * $this->factor;
+            $h = $src_h * $this->factor;
+            $image->crop( 0, 0, $src_w, $src_h, $w, $h );
+            $result = $image->save( $save_filename );
+            if ( is_wp_error( $result ) ) {
+                error_log( 'Error resizing image' );
+                error_log( print_r( $result, true ) );
+                return false;
+            } else {
+                return true;
+            }
+        } else if ( isset( $image->error_data['error_loading_image'] ) ) {
+            TimberHelper::error_log( 'Error loading ' . $image->error_data['error_loading_image'] );
+        } else {
+            TimberHelper::error_log( $image );
+        }
+        return false;
+    }
+}
