+++ functions/timber-image-helper.php
@@ -23,7 +23,7 @@ class TimberImageHelper {
      * @param int     $h target heighth
      * @param string  $crop
      * @param bool    $force_resize
-     * @return string
+     * @return string (ex: )
      */
     public static function resize( $src, $w, $h = 0, $crop = 'default', $force = false ) {
         $op = new TimberImageOperationResize($w, $h, $crop);
@@ -95,6 +95,14 @@ class TimberImageHelper {
             $wp_content_path = str_replace( home_url(), '', WP_CONTENT_URL );
             define( 'WP_CONTENT_SUBDIR', $wp_content_path );
         }
+    }
+
+    static function load_dependencies() {
+        require_once('image/timber-image-operation.php');   
+        require_once('image/timber-image-operation-pngtojpg.php');   
+        require_once('image/timber-image-operation-retina.php');   
+        require_once('image/timber-image-operation-letterbox.php');   
+        require_once('image/timber-image-operation-resize.php');   
     }
 
     /**
@@ -455,318 +463,8 @@ class TimberImageHelper {
 
 }
 
-/**
- * Each image filter is represented by a subclass of this class,m
- * and each filter call is a new instance, with call arguments as properties.
- * 
- * Only 3 methods need to be implemented:
- * - constructor, storing all filter arguments
- * - filename
- * - run
- */
-abstract class TimberImageOperation {
-    /**
-     * Builds the result filename, based on source filename and extension
-     * 
-     * @param  string $src_filename  source filename (excluding extension and path)
-     * @param  string $src_extension source file extension
-     * @return string                resulting filename (including extension but excluding path)
-     */
-    public abstract function filename($src_filename, $src_extension);
 
-    /**
-     * Performs the actual image manipulation,
-     * including saving the target file.
-     * 
-     * @param  string $load_filename filepath (not URL) to source file
-     * @param  string $save_filename filepath (not URL) where result file should be saved
-     * @return bool                  true if everything went fine, false otherwise
-     */
-    public abstract function run($load_filename, $save_filename);
-
-    /**
-     * Helper method to convert hex string to rgb array
-     * 
-     * @param  string $hexstr hex color string (like '#FF1455')
-     * @return array          array('red', 'green', 'blue') to int
-     */
-    public static function hexrgb( $hexstr ) {
-        if ( !strstr( $hexstr, '#' ) ) {
-            $hexstr = '#' . $hexstr;
-        }
-        if ( strlen( $hexstr ) == 4 ) {
-            $hexstr = '#' . $hexstr[1] . $hexstr[1] . $hexstr[2] . $hexstr[2] . $hexstr[3] . $hexstr[3];
-        }
-        $int = hexdec( $hexstr );
-        return array( "red" => 0xFF & ( $int >> 0x10 ), "green" => 0xFF & ( $int >> 0x8 ), "blue" => 0xFF & $int );
-    }
-}
-
-/**
- * Implements converting a PNG file to JPG.
- * Argument:
- * - color to fill transparent zones
- */
-class TimberImageOperationPngToJpg extends TimberImageOperation {
-    private $color;
-
-    /**
-     * @param string $color hex string of color to use for transparent zones
-     */
-    function __construct($color) {
-        $this->color = $color;
-    }
-
-    function filename($src_filename, $src_extension) {
-        $new_name = $src_filename . '.jpg';
-        return $new_name;
-    }
-
-    function run($load_filename, $save_filename){
-        $input = imagecreatefrompng( $load_filename );
-        list( $width, $height ) = getimagesize( $load_filename );
-        $output = imagecreatetruecolor( $width, $height );
-        $c = self::hexrgb( $this->color );
-        $color = imagecolorallocate( $output, $c['red'], $c['green'], $c['blue'] );
-        imagefilledrectangle( $output, 0, 0, $width, $height, $color );
-        imagecopy( $output, $input, 0, 0, 0, 0, $width, $height );
-        imagejpeg( $output, $save_filename );
-        return true;
-    }
-}
-
-/**
- * Increases image size by a given factor
- * Arguments:
- * - factor by which to multiply image dimensions
- */
-class TimberImageOperationRetina extends TimberImageOperation {
-    private $factor;
-
-    /**
-     * @param int $factor to multiply original dimensions by
-     */
-    function __construct($factor) {
-        $this->factor = $factor;
-    }
-
-    function filename($src_filename, $src_extension) {
-        $newbase = $src_filename . '@' . $this->factor . 'x'; // add @2x, @3x, @1.5x, etc.
-        $new_name = $newbase . '.' . $src_extension;
-        return $new_name;
-    }
-
-    function run($load_filename, $save_filename){
-        $image = wp_get_image_editor( $load_filename );
-        if ( !is_wp_error( $image ) ) {
-            $current_size = $image->get_size();
-            $src_w = $current_size['width'];
-            $src_h = $current_size['height'];
-            $src_ratio = $src_w / $src_h;
-            // Get ratios
-            $w = $src_w * $this->factor;
-            $h = $src_h * $this->factor;
-            $image->crop( 0, 0, $src_w, $src_h, $w, $h );
-            $result = $image->save( $save_filename );
-            if ( is_wp_error( $result ) ) {
-                error_log( 'Error resizing image' );
-                error_log( print_r( $result, true ) );
-                return false;
-            } else {
-                return true;
-            }
-        } else if ( isset( $image->error_data['error_loading_image'] ) ) {
-            TimberHelper::error_log( 'Error loading ' . $image->error_data['error_loading_image'] );
-        } else {
-            TimberHelper::error_log( $image );
-        }
-        return false;
-    }
-}
-
-/**
- * Changes image to new size, by shrinking/enlarging then padding with colored bands,
- * so that no part of the image is cropped or stretched.
- * 
- * Arguments:
- * - width of new image
- * - height of new image
- * - color of padding 
- */
-class TimberImageOperationLetterbox extends TimberImageOperation {
-    private $w, $h, $color;
-
-    /**
-     * @param int    $w     width of result image
-     * @param int    $h     height
-     * @param string $color hex string, for color of padding bands
-     */
-    function __construct($w, $h, $color) {
-        $this->w = $w;
-        $this->h = $h;
-        $this->color = $color;
-    }
-
-    public function filename($src_filename, $src_extension) {
-        $color = str_replace( '#', '', $this->color );
-        $newbase = $src_filename . '-lbox-' . $this->w . 'x' . $this->h . '-' . $color;
-        $new_name = $newbase . '.' . $src_extension;
-        return $new_name;
-    }
-
-    public function run($load_filename, $save_filename) {
-        $w = $this->w;
-        $h = $this->h;
-
-        $bg = imagecreatetruecolor( $w, $h );
-        $c = self::hexrgb( $this->color );
-        $bgColor = imagecolorallocate( $bg, $c['red'], $c['green'], $c['blue'] );
-        imagefill( $bg, 0, 0, $bgColor );
-        $image = wp_get_image_editor( $load_filename );
-        if ( !is_wp_error( $image ) ) {
-            $current_size = $image->get_size();
-            $ow = $current_size['width'];
-            $oh = $current_size['height'];
-            $new_aspect = $w / $h;
-            $old_aspect = $ow / $oh;
-            if ( $new_aspect > $old_aspect ) {
-                //taller than goal
-                $h_scale = $h / $oh;
-                $owt = $ow * $h_scale;
-                $y = 0;
-                $x = $w / 2 - $owt / 2;
-                $oht = $h;
-                $image->crop( 0, 0, $ow, $oh, $owt, $oht );
-            } else {
-                $w_scale = $w / $ow;
-                $oht = $oh * $w_scale;
-                $x = 0;
-                $y = $h / 2 - $oht / 2;
-                $owt = $w;
-                $image->crop( 0, 0, $ow, $oh, $owt, $oht );
-            }
-            $image->save( $save_filename );
-            $func = 'imagecreatefromjpeg';
-            $ext = pathinfo( $save_filename, PATHINFO_EXTENSION );
-            if ( $ext == 'gif' ) {
-                $func = 'imagecreatefromgif';
-            } else if ( $ext == 'png' ) {
-                $func = 'imagecreatefrompng';
-            }
-            $image = $func( $save_filename );
-            imagecopy( $bg, $image, $x, $y, 0, 0, $owt, $oht );
-            imagejpeg( $bg, $save_filename );
-            return true;
-        } else {
-            TimberHelper::error_log( $image );
-        }
-        return false;
-    }
-}
-
-/**
- * Changes image to new size, by shrinking/enlarging
- * then cropping to respect new ratio.
- * 
- * Arguments:
- * - width of new image
- * - height of new image
- * - crop method 
- */
-class TimberImageOperationResize extends TimberImageOperation {
-    private $w, $h, $crop;
-
-    /**
-     * @param int    $w    width of new image
-     * @param int    $h    height of new image
-     * @param string $crop cropping method, one of: 'default', 'center', 'top', 'bottom', 'left', 'right'.
-     */
-    function __construct($w, $h, $crop) {
-        $this->w = $w;
-        $this->h = $h;
-        // Sanitize crop position
-        $allowed_crop_positions = array( 'default', 'center', 'top', 'bottom', 'left', 'right' );
-        if ( $crop !== false && !in_array( $crop, $allowed_crop_positions ) ) {
-            $crop = $allowed_crop_positions[0];
-        }
-        $this->crop = $crop;
-    }
-
-    public function filename($src_filename, $src_extension) {
-        $result = $src_filename . '-' . $this->w . 'x' . $this->h . '-c-' . ( $this->crop ? $this->crop : 'f' ); // Crop will be either user named or f (false)
-        if($src_extension) {
-            $result .= '.'.$src_extension;
-        }
-        return $result;
-    }
-
-    public function run($load_filename, $save_filename) {
-        $image = wp_get_image_editor( $load_filename );
-        if ( !is_wp_error( $image ) ) {
-            $w = $this->w;
-            $h = $this->h;
-            $crop = $this->crop;
-
-            $current_size = $image->get_size();
-            $src_w = $current_size['width'];
-            $src_h = $current_size['height'];
-            $src_ratio = $src_w / $src_h;
-            if ( !$h ) {
-                $h = round( $w / $src_ratio );
-            }
-            if ( !$w ) {
-                //the user wants to resize based on constant height
-                $w = round( $h * $src_ratio );
-            }
-            // Get ratios
-            $dest_ratio = $w / $h;
-            $src_wt = $src_h * $dest_ratio;
-            $src_ht = $src_w / $dest_ratio;
-            if ( !$crop ) {
-                // Always crop, to allow resizing upwards
-                $image->crop( 0, 0, $src_w, $src_h, $w, $h );
-            } else {
-                //start with defaults:
-                $src_x = $src_w / 2 - $src_wt / 2;
-                $src_y = ( $src_h - $src_ht ) / 6;
-                //now specific overrides based on options:
-                if ( $crop == 'center' ) {
-                    // Get source x and y
-                    $src_x = round( ( $src_w - $src_wt ) / 2 );
-                    $src_y = round( ( $src_h - $src_ht ) / 2 );
-                } else if ( $crop == 'top' ) {
-                        $src_y = 0;
-                    } else if ( $crop == 'bottom' ) {
-                        $src_y = $src_h - $src_ht;
-                    } else if ( $crop == 'left' ) {
-                        $src_x = 0;
-                    } else if ( $crop == 'right' ) {
-                        $src_x = $src_w - $src_wt;
-                    }
-                // Crop the image
-                if ( $dest_ratio > $src_ratio ) {
-                    $image->crop( 0, $src_y, $src_w, $src_ht, $w, $h );
-                } else {
-                    $image->crop( $src_x, 0, $src_wt, $src_h, $w, $h );
-                }
-            }
-            $result = $image->save( $save_filename );
-            if ( is_wp_error( $result ) ) {
-                error_log( 'Error resizing image' );
-                error_log( print_r( $result, true ) );
-                return false;
-            } else {
-                return true;
-            }
-        } else if ( isset( $image->error_data['error_loading_image'] ) ) {
-            TimberHelper::error_log( 'Error loading ' . $image->error_data['error_loading_image'] );
-        } else {
-            TimberHelper::error_log( $image );
-        }
-        return false;
-    }
-}
-
+TimberImageHelper::load_dependencies();
 TimberImageHelper::add_constants();
 TimberImageHelper::add_actions();
 TimberImageHelper::add_filters();
